build: update RN files and change project name
