+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "OwlrangeNotes",
-  "displayName": "OwlrangeNotes"
+  "name": "StudynON",
+  "displayName": "Studyn'ON"
 }
