+++ ios/StudynON.xcodeproj/xcshareddata/xcschemes/StudynON.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "OwlrangeNotes.app"
-               BlueprintName = "OwlrangeNotes"
-               ReferencedContainer = "container:OwlrangeNotes.xcodeproj">
+               BuildableName = "StudynON.app"
+               BlueprintName = "StudynON"
+               ReferencedContainer = "container:StudynON.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "OwlrangeNotesTests.xctest"
-               BlueprintName = "OwlrangeNotesTests"
-               ReferencedContainer = "container:OwlrangeNotes.xcodeproj">
+               BuildableName = "StudynONTests.xctest"
+               BlueprintName = "StudynONTests"
+               ReferencedContainer = "container:StudynON.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "OwlrangeNotes.app"
-            BlueprintName = "OwlrangeNotes"
-            ReferencedContainer = "container:OwlrangeNotes.xcodeproj">
+            BuildableName = "StudynON.app"
+            BlueprintName = "StudynON"
+            ReferencedContainer = "container:StudynON.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "OwlrangeNotes.app"
-            BlueprintName = "OwlrangeNotes"
-            ReferencedContainer = "container:OwlrangeNotes.xcodeproj">
+            BuildableName = "StudynON.app"
+            BlueprintName = "StudynON"
+            ReferencedContainer = "container:StudynON.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
