+++ ios/StudynON/Images.xcassets/AppIcon.appiconset/Contents.json
