+++ ios/StudynONTests/Info.plist
