+++ ios/StudynONTests/StudynONTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface OwlrangeNotesTests : XCTestCase
+@interface StudynONTests : XCTestCase
 
 @end
 
-@implementation OwlrangeNotesTests
+@implementation StudynONTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
